// src/components/your-stone/your-stone.types.ts
export interface Settings {
  name: string;
  title: string;
  subtitle: string;
  view_all: string;
  view_all_link: string;
  filter_stone: '1' | '2' | '3' | '4' | '5';
  title_1: string;
  image_1: string | null;
  title_2: string;
  image_2: string | null;
  title_3: string;
  image_3: string | null;
  title_4: string;
  image_4: string | null;
  title_5: string;
  image_5: string | null;
}